+++ test/unit/modules/api/_helper.ts
@@ -5,7 +5,7 @@ import supertest from 'supertest';
 
 import { parseConfigFile } from '@verdaccio/config';
 import { HEADERS, HEADER_TYPE, HTTP_STATUS, TOKEN_BEARER } from '@verdaccio/core';
-import { GenericBody, PackageUsers } from '@verdaccio/types';
+import { GenericBody, PackageUsers, ConfigYaml } from '@verdaccio/types';
 import { buildToken, generateRandomHexString } from '@verdaccio/utils';
 
 import apiMiddleware from '../../../../src/api/endpoint';
@@ -24,8 +24,12 @@ export const getConf = (conf) => {
   return config;
 };
 
-export async function initializeServer(configName): Promise<Application> {
+export async function initializeServer(configName: string): Promise<Application> {
   const config = getConf(configName);
+  return initializeServerHelper(config, [apiMiddleware], Storage);
+}
+
+export async function initializeServerWithConfig(config: ConfigYaml): Promise<Application> {  
   return initializeServerHelper(config, [apiMiddleware], Storage);
 }
